backend/requirements.txt
# MAXY Chat Backend - Python Dependencies

# Core Framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0

# File Processing - Use compatible versions for Python 3.14
Pillow>=10.2.0
PyPDF2>=3.0.0
python-docx>=0.8.11

# Data Analysis & Processing - Flexible versions for compatibility
numpy>=1.26.0
scipy>=1.11.0
pandas>=2.1.0

# Visualization
matplotlib>=3.8.0
plotly>=5.18.0

# Web & Research
requests>=2.31.0
wikipedia>=1.4.0

# Utilities
python-multipart>=0.0.6
aiofiles>=23.2.0

# Logging & Monitoring
structlog>=23.2.0

# Production WSGI
gunicorn>=21.0.0

# PostgreSQL (for production)
# psycopg2-binary>=2.9.9

# Development (optional)
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# black>=23.0.0
# flake8>=6.0.0
